:recycle: Presenter로 Theater, MovieInfo 데이터 분리
